requirements/base.txt
# SHA1:f68f28352f1293b1d6557f881cdd366cbc5bd919
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
confluent-kafka==2.6.2
    # via -r base.in
contourpy==1.3.1
    # via matplotlib
cyclebane==24.10.0
    # via sciline
cycler==0.12.1
    # via matplotlib
ess-streaming-data-types==0.26.1
    # via -r base.in
essreduce==24.11.3
    # via -r base.in
flatbuffers==24.3.25
    # via ess-streaming-data-types
fonttools==4.55.3
    # via matplotlib
h5py==3.12.1
    # via
    #   scippneutron
    #   scippnexus
jinja2==3.1.4
    # via -r base.in
kiwisolver==1.4.7
    # via matplotlib
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.0
    # via
    #   mpltoolbox
    #   plopp
mpltoolbox==24.5.1
    # via scippneutron
networkx==3.4.2
    # via cyclebane
numpy==2.2.0
    # via
    #   contourpy
    #   ess-streaming-data-types
    #   h5py
    #   matplotlib
    #   mpltoolbox
    #   scipp
    #   scippneutron
    #   scipy
packaging==24.2
    # via matplotlib
pillow==11.0.0
    # via matplotlib
plopp==24.10.0
    # via scippneutron
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   scippnexus
pyyaml==6.0.2
    # via -r base.in
sciline==24.10.0
    # via essreduce
scipp==24.11.2
    # via
    #   essreduce
    #   scippneutron
    #   scippnexus
scippneutron==24.12.0
    # via essreduce
scippnexus==24.11.1
    # via
    #   essreduce
    #   scippneutron
scipy==1.14.1
    # via
    #   scippneutron
    #   scippnexus
six==1.17.0
    # via python-dateutil
